[wdspec] rename `emulation.setClientHintsOverride` -> `userAgentClientHints.setClientHintsOverride`
Follow-up after https://github.com/WICG/ua-client-hints/pull/393
