+++ webdriver/tests/bidi/external/user_agent_client_hints/set_client_hints_override/contexts.py
@@ -8,7 +8,7 @@ pytestmark = pytest.mark.asyncio
 async def test_contexts(bidi_session, new_tab, top_context,
         default_client_hints, assert_client_hints):
     # Set client hints override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         contexts=[new_tab["context"]],
         client_hints=SOME_CLIENT_HINTS
     )
@@ -18,7 +18,7 @@ async def test_contexts(bidi_session, new_tab, top_context,
     await assert_client_hints(top_context, default_client_hints)
 
     # Reset client hints override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         contexts=[new_tab["context"]],
         client_hints=None
     )
@@ -33,7 +33,7 @@ async def test_multiple_contexts(bidi_session, new_tab, default_client_hints,
     new_context = await bidi_session.browsing_context.create(type_hint="tab")
 
     # Set client hints override
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         contexts=[new_tab["context"], new_context["context"]],
         client_hints=SOME_CLIENT_HINTS
     )
@@ -43,7 +43,7 @@ async def test_multiple_contexts(bidi_session, new_tab, default_client_hints,
     await assert_client_hints(new_context, SOME_CLIENT_HINTS)
 
     # Reset client hints override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         contexts=[new_tab["context"], new_context["context"]],
         client_hints=None
     )
