+++ webdriver/tests/bidi/external/user_agent_client_hints/set_client_hints_override/global.py
@@ -14,7 +14,7 @@ async def test_set_override_and_reset_globally(bidi_session,
     context_in_user_context = await bidi_session.browsing_context.create(
         user_context=user_context, type_hint="tab")
 
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         client_hints=SOME_CLIENT_HINTS
     )
 
@@ -28,7 +28,7 @@ async def test_set_override_and_reset_globally(bidi_session,
                               SOME_CLIENT_HINTS)
 
     # Reset global override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         client_hints=None
     )
 
@@ -46,14 +46,14 @@ async def test_set_override_and_reset_globally(bidi_session,
 
 async def test_set_override_and_reset_globally_and_per_context(
         bidi_session, top_context, default_client_hints, assert_client_hints):
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         contexts=[top_context["context"]],
         client_hints=SOME_CLIENT_HINTS
     )
     await assert_client_hints(top_context, SOME_CLIENT_HINTS)
 
     # Set global override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         client_hints=ANOTHER_CLIENT_HINTS
     )
 
@@ -61,7 +61,7 @@ async def test_set_override_and_reset_globally_and_per_context(
     await assert_client_hints(top_context, SOME_CLIENT_HINTS)
 
     # Reset per-context override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         contexts=[top_context["context"]],
         client_hints=None
     )
@@ -70,7 +70,7 @@ async def test_set_override_and_reset_globally_and_per_context(
     await assert_client_hints(top_context, ANOTHER_CLIENT_HINTS)
 
     # Reset global override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         client_hints=None
     )
 
@@ -80,14 +80,14 @@ async def test_set_override_and_reset_globally_and_per_context(
 
 async def test_set_override_and_reset_globally_and_per_user_context(
         bidi_session, top_context, default_client_hints, assert_client_hints):
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         user_contexts=["default"],
         client_hints=SOME_CLIENT_HINTS
     )
     await assert_client_hints(top_context, SOME_CLIENT_HINTS)
 
     # Set global override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         client_hints=ANOTHER_CLIENT_HINTS
     )
 
@@ -95,7 +95,7 @@ async def test_set_override_and_reset_globally_and_per_user_context(
     await assert_client_hints(top_context, SOME_CLIENT_HINTS)
 
     # Reset per user context override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         user_contexts=["default"],
         client_hints=None
     )
@@ -104,7 +104,7 @@ async def test_set_override_and_reset_globally_and_per_user_context(
     await assert_client_hints(top_context, ANOTHER_CLIENT_HINTS)
 
     # Reset global override.
-    await bidi_session.emulation.set_client_hints_override(
+    await bidi_session.user_agent_client_hints.set_client_hints_override(
         client_hints=None
     )
