+++ webdriver/tests/bidi/external/user_agent_client_hints/set_client_hints_override/invalid.py
@@ -12,7 +12,7 @@ pytestmark = pytest.mark.asyncio
 @pytest.mark.parametrize("value", get_invalid_cases("list"))
 async def test_params_contexts_invalid_type(bidi_session, value):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             contexts=value
         )
@@ -20,7 +20,7 @@ async def test_params_contexts_invalid_type(bidi_session, value):
 
 async def test_params_contexts_empty_list(bidi_session):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             contexts=[])
 
@@ -28,14 +28,14 @@ async def test_params_contexts_empty_list(bidi_session):
 @pytest.mark.parametrize("value", get_invalid_cases("string"))
 async def test_params_contexts_entry_invalid_type(bidi_session, value):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             contexts=[value])
 
 
 async def test_params_contexts_entry_invalid_value(bidi_session):
     with pytest.raises(error.NoSuchFrameException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             contexts=["_invalid_"],
         )
@@ -54,7 +54,7 @@ async def test_params_contexts_iframe(bidi_session, new_tab, get_test_page):
     assert len(frames) == 1
 
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             contexts=[frames[0]["context"]],
         )
@@ -63,7 +63,7 @@ async def test_params_contexts_iframe(bidi_session, new_tab, get_test_page):
 @pytest.mark.parametrize("value", get_invalid_cases("list"))
 async def test_params_user_contexts_invalid_type(bidi_session, value):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             user_contexts=value,
         )
@@ -71,7 +71,7 @@ async def test_params_user_contexts_invalid_type(bidi_session, value):
 
 async def test_params_user_contexts_empty_list(bidi_session):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             user_contexts=[],
         )
@@ -80,7 +80,7 @@ async def test_params_user_contexts_empty_list(bidi_session):
 @pytest.mark.parametrize("value", get_invalid_cases("string"))
 async def test_params_user_contexts_entry_invalid_type(bidi_session, value):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             user_contexts=[value],
         )
@@ -89,7 +89,7 @@ async def test_params_user_contexts_entry_invalid_type(bidi_session, value):
 @pytest.mark.parametrize("value", ["", "somestring"])
 async def test_params_user_contexts_entry_invalid_value(bidi_session, value):
     with pytest.raises(error.NoSuchUserContextException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             user_contexts=[value],
         )
@@ -97,7 +97,7 @@ async def test_params_user_contexts_entry_invalid_value(bidi_session, value):
 
 async def test_params_contexts_and_user_contexts(bidi_session, top_context):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=SOME_CLIENT_HINTS,
             contexts=[top_context["context"]],
             user_contexts=["default"],
@@ -106,7 +106,7 @@ async def test_params_contexts_and_user_contexts(bidi_session, top_context):
 
 async def test_params_client_hints_missing(bidi_session, top_context):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=UNDEFINED,
             contexts=[top_context["context"]],
         )
@@ -116,7 +116,7 @@ async def test_params_client_hints_missing(bidi_session, top_context):
 async def test_params_client_hints_invalid_type(bidi_session, top_context,
         value):
     with pytest.raises(error.InvalidArgumentException):
-        await bidi_session.emulation.set_client_hints_override(
+        await bidi_session.user_agent_client_hints.set_client_hints_override(
             client_hints=value,
             contexts=[top_context["context"]],
         )
